requirements.txt
asyncio
beautifulsoup4
playwright
pandas
python-dotenv
openpyxl

